requirements.txt
asgiref==3.8.1
attrs==25.1.0
beautifulsoup4==4.14.2
bleach==6.2.0
boolean.py==5.0
Brotli==1.1.0
CacheControl==0.14.3
certifi==2025.10.5
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.4
coverage==7.7.1
factory_boy==3.3.0
cssselect2==0.8.0
cyclonedx-python-lib==9.1.0
defusedxml==0.7.1
Django==4.2.7
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
django-guardian==2.4.0
django-jazzmin==3.0.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.5.1
et_xmlfile==2.0.0
filelock==3.20.0
fonttools==4.58.4
idna==3.11
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
license-expression==30.4.4
lxml==5.3.1
markdown-it-py==4.0.0
mdurl==0.1.2
msgpack==1.1.2
numpy==2.2.3
openpyxl==3.1.5
packageurl-python==0.17.5
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pdfrw==0.4
pillow==11.1.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.5.0
pluggy==1.5.0
psycopg2-binary==2.9.10
py-serializable==2.1.0
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.2
PyJWT==2.9.0
PyMuPDF==1.26.5
pyparsing==3.2.5
PyPDF2==3.0.1
pyphen==0.17.2
pytesseract==0.3.13
pytest==8.3.5
pytest-django==4.10.0
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.1
python-magic-bin==0.4.14
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.3.1
requests==2.32.5
rich==14.2.0
rpds-py==0.23.1
setuptools==80.9.0
six==1.17.0
sortedcontainers==2.4.0
soupsieve==2.8
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.5.0
watchdog==6.0.0
weasyprint==65.1
webencodings==0.5.1
wheel==0.45.1
whitenoise==6.6.0
zopfli==0.2.3.post1
